+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.6] - 2025-03-18 ([#13](https://github.com/lucas-soria/qrsec_backend/pull/13))
+### Removed
+- Removed password from User entity.
+- Removed password encoder service.
+
+### Added
+- Added "check if user exists" endpoint.
+
 ## [0.0.5] - 2025-01-22 ([#11](https://github.com/lucas-soria/qrsec_backend/pull/11))
 ### Removed
 - /admin/ endpoints.